fix issue with useSearchParams
